Exclude serial transport into separate assembly
