+++ NModbus4.sln
@@ -10,6 +10,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		global.json = global.json
 	EndProjectSection
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NModbus4.UnitTests", "NModbus4.UnitTests\NModbus4.UnitTests.xproj", "{099155AF-8348-4829-B959-2F83AFCDD4E3}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "NModbus4.Serial", "NModbus4.Serial\NModbus4.Serial.xproj", "{2246217D-D3C2-41CB-B866-D093DF087846}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -20,6 +24,14 @@ Global
 		{9B07ED72-3AD9-45EC-AAE9-295849B309AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9B07ED72-3AD9-45EC-AAE9-295849B309AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9B07ED72-3AD9-45EC-AAE9-295849B309AA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{099155AF-8348-4829-B959-2F83AFCDD4E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{099155AF-8348-4829-B959-2F83AFCDD4E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{099155AF-8348-4829-B959-2F83AFCDD4E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{099155AF-8348-4829-B959-2F83AFCDD4E3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2246217D-D3C2-41CB-B866-D093DF087846}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2246217D-D3C2-41CB-B866-D093DF087846}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2246217D-D3C2-41CB-B866-D093DF087846}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2246217D-D3C2-41CB-B866-D093DF087846}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
